refactor(merk/proofs): split aggregate_sum.rs into aggregate_sum/ subdirectory
The single-file `aggregate_sum.rs` had grown to 1557 lines, mixing six
concerns: module docs, type predicates, the public `impl RefWalker`
entry points, the recursive proof emitter, the no-proof walker, the
verifier (entry point + recursive shape walk), and a 753-line tests
module. Each was self-contained but they shared a single namespace,
which made navigation and review unnecessarily hard.

Split along the natural seams:

| File             | Lines | Contents                                |
|------------------|-------|-----------------------------------------|
| `mod.rs`         |   90  | Module docs, layout doc-comment, public re-export of `verify_aggregate_sum_on_range_proof`, the two small tree-type helpers `is_provable_sum_bearing` / `provable_sum_from_aggregate` |
| `prove.rs`       |  127  | `impl RefWalker { create_aggregate_sum_on_range_proof, sum_aggregate_on_range }` |
| `emit.rs`        |  265  | `emit_sum_proof` — the recursive proof-emission engine |
| `walk.rs`        |  179  | `walk_sum_only` + `provable_sum_from_walker` (no-proof variant) |
| `verify.rs`      |  256  | `verify_aggregate_sum_on_range_proof` + `verify_sum_shape` |
| `tests.rs`       |  771  | All unit + integration + fuzz tests    |

Total: ~1688 lines across 6 files (was 1557 in one) — the small
size growth is from the per-file `//!` doc headers and the explicit
import lists tests.rs now needs (the old single-file `use super::*;`
implicitly pulled in everything).

Behavior is identical:
- `cargo test --workspace`: 3199 / 0 fail (unchanged).
- `cargo clippy --workspace --all-features`: clean.
- The public surface (`Merk::prove_aggregate_sum_on_range`,
  `Merk::sum_aggregate_on_range`, `verify_aggregate_sum_on_range_proof`)
  is untouched — the only externally-visible re-export in
  `merk::proofs::query::mod.rs` (`pub use aggregate_sum::verify_aggregate_sum_on_range_proof`)
  still resolves correctly because Rust treats `module.rs` and
  `module/mod.rs` as interchangeable.

The companion `aggregate_count.rs` (2018 lines) is structurally
analogous and would benefit from the same split. Left untouched here
to keep this PR focused on aggregate_sum (the user's request); the
count-side restructure is a natural follow-up.
